Move calculator class to a seperate file
